refactor: 키워드가 공란일 때 전체 조회되도록 수정

+++ src/main/java/com/dangun/miniproject/board/dto/WriteBoardRequest.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 public class WriteBoardRequest {
     private final String title;
     private final String content;
-    private final Long price;
+    private final Integer price;
 }
